Add files via upload
his script allows the user to input a maximum number to start the game. Then, the user makes guesses to find the random number generated by the computer. If the guess is below or above the random number, an output is printed stating the same. If the user guesses the correct number, the game ends, and the total number of guesses made by the user is displayed.