Add files via upload
A simple Rock-Paper-Scissors game build with python. User can input rock/paper/scissors and input of computer is randomized by the random.randint function. Both inputs are compared, and You lost and You won are shows based on the criteria. If user decides to quit, they can press Q to do so. After breaking of operation, a score is showed for both user and the computer.